Limit number of datapoints in dataset
A feature is added to downsample each simulation set to contain a max number of datapoints. This was needed to improve interactive graph performance. This was done by creating a function 'downsampleList'. The threshold has been set to 500 by default, but can be modified by the user by changing the add-on configurations.